Export and import fixes
