+++ src/components/SearchBox/index.js
@@ -1,0 +1,3 @@
+import SearchBox from "./SearchBox";
+
+export default SearchBox;
